Remove tech trees from FactionSummary.
