+++ src/FactionSummary.module.scss
@@ -1,6 +1,6 @@
 .FactionSummary {
   display: flex;
-  gap: 0.375rem;
+  gap: 1rem;
   justify-content: space-between;
   padding: 0 0 0 0.25rem;
   position: relative;
@@ -15,56 +15,8 @@
   gap: 0;
   font-size: 1.75rem;
   justify-items: center;
-  padding-right: 0.375rem;
 }
 
 .VPGrid :last-child {
   grid-column: 1 / 4;
-}
-
-.TechSummaryGrid {
-  display: grid;
-  font-size: 0.875rem;
-  row-gap: 0.125rem;
-  column-gap: 0.0625rem;
-  grid-template-columns: repeat(8, 1fr);
-  width: min-content;
-}
-
-.TechSummaryGrid.Horizontal {
-  grid-template-columns: repeat(15, 1fr);
-  font-size: 1.125rem;
-  row-gap: 0.1875rem;
-  column-gap: 0.1875rem;
-}
-
-.TechSummaryGrid.Horizontal > .UnitUpgradeNumber {
-  grid-column: 5 / 6;
-}
-
-.TechSummaryGrid > .UnitUpgradeText {
-  grid-column: 1 / 8;
-}
-
-.TechSummaryGrid.Horizontal > .UnitUpgradeText {
-  grid-column: 5 / 13;
-}
-
-.TechSummaryGrid > .UnitUpgradeTree {
-  grid-column: 1 / 8;
-}
-
-.TechSummaryGrid.Horizontal > .UnitUpgradeTree {
-  grid-column: 5 / 13;
-}
-
-.TechSummaryGrid > .FactionTechTree {
-  padding-left: 0.375rem;
-  grid-column: 8 / 9;
-  grid-row: 1 / 5;
-}
-
-.TechSummaryGrid.Horizontal > .FactionTechTree {
-  grid-column: 16 / 17;
-  grid-row: 1 / 4;
 }
