+++ src/FactionSummary.tsx
@@ -3,7 +3,6 @@ import { FormattedMessage } from "react-intl";
 import styles from "./FactionSummary.module.scss";
 import FactionIcon from "./components/FactionIcon/FactionIcon";
 import PlanetSummary from "./components/PlanetSummary/PlanetSummary";
-import TechIcon from "./components/TechIcon/TechIcon";
 import { GameIdContext } from "./context/Context";
 import {
   useActionLog,
@@ -20,144 +19,14 @@ import {
   scoreObjectiveAsync,
   unscoreObjectiveAsync,
 } from "./dynamic/api";
-import { computeScoredVPs, getFactionColor } from "./util/factions";
+import { computeScoredVPs } from "./util/factions";
 import {
   applyAllPlanetAttachments,
   filterToClaimedPlanets,
 } from "./util/planets";
 import { filterToOwnedTechs } from "./util/techs";
-import TechTree from "./components/TechTree/TechTree";
 import { objectEntries, rem } from "./util/util";
-
-export function TechSummary({
-  techs,
-  factionId,
-  horizontal,
-}: {
-  techs: Tech[];
-  factionId: FactionId;
-  horizontal?: boolean;
-}) {
-  let blueTechs = [];
-  let yellowTechs = [];
-  let greenTechs = [];
-  let redTechs = [];
-  let upgradeTechs = [];
-  for (const tech of techs) {
-    switch (tech.type) {
-      case "RED":
-        redTechs.push(tech);
-        break;
-      case "YELLOW":
-        yellowTechs.push(tech);
-        break;
-      case "GREEN":
-        greenTechs.push(tech);
-        break;
-      case "BLUE":
-        blueTechs.push(tech);
-        break;
-      case "UPGRADE":
-        upgradeTechs.push(tech);
-        break;
-    }
-  }
-
-  const techOrder: TechType[] = ["GREEN", "BLUE", "YELLOW", "RED", "UPGRADE"];
-
-  techs.sort((a, b) => {
-    const typeDiff = techOrder.indexOf(a.type) - techOrder.indexOf(b.type);
-    if (typeDiff !== 0) {
-      return typeDiff;
-    }
-    const prereqDiff: number = a.prereqs.length - b.prereqs.length;
-    if (prereqDiff !== 0) {
-      return prereqDiff;
-    }
-    if (a.name < b.name) {
-      return -1;
-    } else {
-      return 1;
-    }
-  });
-
-  const numberWidth = horizontal ? rem(10.75 * 1.2) : rem(10.75);
-  const techTreeSize = horizontal ? 6 : 4;
-  const iconSize = horizontal ? 20 : 16;
-
-  return (
-    <>
-      <div
-        className={`${styles.TechSummaryGrid} ${
-          horizontal ? styles.Horizontal : ""
-        }`}
-      >
-        <div
-          className="flexRow centered"
-          style={{ height: "100%", width: numberWidth }}
-        >
-          {greenTechs.length || "-"}
-        </div>
-        <div className="flexRow" style={{ height: "100%" }}>
-          <TechIcon type={"GREEN"} size={iconSize} />
-        </div>
-        <TechTree type="GREEN" factionId={factionId} size={techTreeSize} />
-        <div>&nbsp;</div>
-        <div
-          className="flexRow centered"
-          style={{ height: "100%", width: numberWidth }}
-        >
-          {blueTechs.length || "-"}
-        </div>
-
-        <div className="flexRow" style={{ height: "100%" }}>
-          <TechIcon type={"BLUE"} size={iconSize} />
-        </div>
-        <TechTree type="BLUE" factionId={factionId} size={techTreeSize} />
-        {horizontal ? <div>&nbsp;</div> : null}
-        <div
-          className="flexRow centered"
-          style={{ height: "100%", width: numberWidth }}
-        >
-          {yellowTechs.length || "-"}
-        </div>
-        <div className="flexRow" style={{ height: "100%" }}>
-          <TechIcon type={"YELLOW"} size={iconSize} />
-        </div>
-        <TechTree type="YELLOW" factionId={factionId} size={techTreeSize} />
-        <div>&nbsp;</div>
-        <div
-          className="flexRow centered"
-          style={{ height: "100%", width: numberWidth }}
-        >
-          {redTechs.length || "-"}
-        </div>
-        <div className="flexRow" style={{ height: "100%" }}>
-          <TechIcon type={"RED"} size={iconSize} />
-        </div>
-        <TechTree type="RED" factionId={factionId} size={techTreeSize} />
-        <div
-          className={styles.UnitUpgradeText}
-          style={{ whiteSpace: "nowrap" }}
-        >
-          {upgradeTechs.length || "-"}{" "}
-          <FormattedMessage
-            id="lGDH2d"
-            description="Unit upgrade techs."
-            defaultMessage="{count, plural, =0 {Upgrades} one {Upgrade} other {Upgrades}}"
-            values={{ count: upgradeTechs.length }}
-          />
-        </div>
-        <div className={styles.UnitUpgradeTree}>
-          <TechTree type="UPGRADE" factionId={factionId} size={techTreeSize} />
-        </div>
-        <div className={styles.FactionTechTree}>
-          <TechTree type="FACTION" factionId={factionId} size={techTreeSize} />
-        </div>
-      </div>
-    </>
-  );
-}
+import TechSummary from "./components/TechSummary/TechSummary";
 
 interface FactionSummaryProps {
   factionId: FactionId;
@@ -213,9 +82,7 @@ export function FactionSummary({
 
   return (
     <div className={styles.FactionSummary}>
-      {options.hideTechs ? null : (
-        <TechSummary techs={ownedTechs} factionId={factionId} />
-      )}
+      {options.hideTechs ? null : <TechSummary techs={ownedTechs} />}
       <div className={styles.VPGrid}>
         {options.showIcon ? (
           <div
